Rename Rewrite module to Mutator
